% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfica.multicore.R
\name{tfica.multicore}
\alias{tfica.multicore}
\title{tfica with multicore computation to save time}
\usage{
tfica.multicore(
  Xin,
  M = dim(Xin)[1],
  Win = diag(M),
  tol = 1e-04,
  maxit = 20,
  nmaxit = 5,
  sopoc = 0.1,
  ncores = 4,
  maxknots = 10
)
}
\arguments{
\item{tol}{(=0.001): convergence criteria}

\item{sopoc}{: Sub-octaves per octave calculated.}

\item{ncores}{(=4)}

\item{Xin:}{M x T data input.}

\item{M:}{number of independent components}

\item{Win:}{initial value for unmixing matrix}

\item{maxit:}{maximum iteration}

\item{nmaxit:}{maximum number of restart}
}
\value{

}
\description{
tfica with multicore computation to save time
}
