% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfplot.R
\name{tfplot}
\alias{tfplot}
\title{Draw 3-panel figure including raw time series, time-frequency power, and average spectral density.}
\usage{
tfplot(
  obj,
  interpolate = TRUE,
  figstr = "fig.png",
  fig.title = "",
  fig.width = 1800,
  fig.height = 1800,
  fig.res = 300
)
}
\arguments{
\item{obj}{: morlet wavelet object (from dplR morlet function)}

\item{fig.res}{}
}
\value{

}
\description{
Draw 3-panel figure including raw time series, time-frequency power, and average spectral density.
}
